discord-price-tracker/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.2
cloudscraper>=1.2.71
playwright>=1.40.0
lxml>=4.9.3
asyncio>=3.4.3